must limit the value of the r1r2 dot product, based on the limit of the pitch circle center to gear tooth tip/edge point, in the XY gear c-sys. (+8 squashed commit)
Squashed commit:

[0ea8fe0] Test code cleaned.

[46a6fb7] stripped debug infos

[750cb51] By default the ANCF node of (position, direction) type should have no atomic mass or inertia. Previously it had non-zero values.
NOTE: in order to write an implementation that could be tested in short time, there is a sort of a hack: the current ChNodeFEAxyzD node  embeds a ChLcpVariablesBodyOwnMass interface to solver (the one that was developed for the rigid bodies). Yes, that is a 6-dof variable as needed, but the mass&inertia info stored in that ChLcpVariablesBodyOwnMass  is overkill for the ANCF node (and, moreover, it is also not well suited to use it in the ANCF context, where the D info is not a rotation info like in rigid bodies).

[dc7bac8] Update to reflect changes in signature of ChFrame::GetA()

[0a558f6] Modified code for testing the beams.
Still there is something flawed with the ANCF beam: the stiffness matrices are correct, yet the displacement is underestimated - see the static test in this source. Will be fixed asap.

[a96b550] The stiffness matrix of the ANCF beam is computed by numerical differentiation. It is less efficient than with analytical formula, but the analytical formula from D.Melanz thesis suffers from a singularity for straight beam with aligned nodes.

The ANCF beam stiffness matrix and internal forces have been validated.

Also, some optimization in the ANCF code has been done.

[58f6de0] MEthod to set the beam rest length (by default it is computed at SetupInitial assuming all nodes are in rest position, but it might be useful otherwise)

[cba9278] gear/pin collision callback working.